enhance: localfs automatically creates folders in writer
In the storage system, there are both `local` and
`remote` filesystems. For remote filesystems, when
performing operations such as PutObject or other
Put actions, S3 automatically creates the parent
directory. However, for the local filesystems, the
parent path is not automatically created.

This commit checks whether the filesystem is local
or remote. If it is a local filesystem, the parent
path will be automatically created during write
operations.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>